+++ src/MailCatcher.php
@@ -3,7 +3,7 @@
 namespace TJVB\MailCatchall;
 
 use Illuminate\Mail\Events\MessageSending;
-use Illuminate\Support\Facades\Log;
+use Psr\Log\LoggerInterface;
 
 /**
  * The class to catch the mail
@@ -12,6 +12,17 @@ use Illuminate\Support\Facades\Log;
  */
 class MailCatcher
 {
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    public function __construct(LoggerInterface $logger)
+    {
+        $this->logger = $logger;
+    }
+
     /**
      * Handle the event.
      *
@@ -29,7 +40,7 @@ class MailCatcher
 
         if (!$receiver) {
             // there isn't a catch all adres configurated so we don't need to do anything
-            Log::error('We can\'t send the mail because the mailcatchall.receiver config value isn\'t set');
+            $this->logger->error('We can\'t send the mail because the mailcatchall.receiver config value isn\'t set');
             return;
         }
         $originalReceivers = [